Set test start_time_ms() to non-zero value. In so doing, I found at least one bug (int instead of int64 used for time).
Make idle flush time configurable. (morlovich)

	Inject fetches at proxy_fetch level when the input connection
	is idle for 10ms. This is in particular helpful for getting the first
	congestion window worth of content out quickly when it doesn't
	have enough content for the flush_html_filter to suggest it's
	worth flusing ASAP. (morlovich)

+++ src/net/instaweb/rewriter/css_image_rewriter.cc
@@ -159,7 +159,7 @@ int64 CssImageRewriter::ExpirationTimeMs(CachedResult* cached_result) {
     // but will likely be able to do so shortly, so we want to expire
     // "almost immediately". We use 1 second as it's the smallest TTL
     // we can reliably represent, anyway.
-    int now_ms = driver_->resource_manager()->timer()->NowMs();
+    int64 now_ms = driver_->resource_manager()->timer()->NowMs();
     return now_ms + Timer::kSecondMs;
   } else {
     return cached_result->origin_expiration_time_ms();
